Move Scene and related types to dedicated module
