+++ lib/render/src/lib.rs
@@ -6,6 +6,7 @@ use math::{Angle, Linear};
 use math::mat::Mat4;
 use math::transform::*;
 use math::vec::*;
+use scene::{Obj, Scene};
 pub use stats::Stats;
 use util::Buffer;
 use util::color::Color;
@@ -15,6 +16,7 @@ use crate::raster::*;
 
 mod hsr;
 pub mod raster;
+pub mod scene;
 pub mod shade;
 pub mod stats;
 pub mod tex;
@@ -41,7 +43,6 @@ pub trait RasterOps<VA: Copy, FA> {
             true
         } else { false }
     }
-
 }
 
 pub struct Raster<Shade, Test, Output> {
@@ -66,56 +67,6 @@ where
     }
     fn output(&mut self, coord: (usize, usize), c: Color) {
         (self.output)(coord, c);
-    }
-}
-
-#[derive(Default, Clone)]
-pub struct Obj<VA, FA> {
-    pub tf: Mat4,
-    pub mesh: Mesh<VA, FA>,
-}
-
-#[derive(Default, Clone)]
-pub struct Scene<VA, FA> {
-    pub objects: Vec<Obj<VA, FA>>,
-    pub camera: Mat4,
-}
-
-
-#[derive(Clone, Debug, Default)]
-pub struct FpsCamera {
-    pub pos: Vec4,
-    pub azimuth: Angle,
-    pub altitude: Angle,
-}
-
-impl FpsCamera {
-    pub fn new(pos: Vec4, azimuth: Angle) -> Self {
-        Self { pos, azimuth, ..Self::default() }
-    }
-
-    pub fn translate(&mut self, dir: Vec4) {
-        let fwd = &rotate_y(self.azimuth) * Z;
-        let right = Y.cross(fwd);
-        self.pos += Vec4::lincomb(fwd, dir.z, right, dir.x);
-    }
-
-    pub fn rotate(&mut self, az: Angle, alt: Angle) {
-        self.azimuth = (self.azimuth + az)
-            .wrap(-Angle::STRAIGHT, Angle::STRAIGHT);
-        self.altitude = (self.altitude + alt)
-            .clamp(-Angle::RIGHT, Angle::RIGHT);
-    }
-
-    pub fn world_to_view(&self) -> Mat4 {
-        let fwd_move = polar(1.0, self.azimuth);
-        let fwd = spherical(1.0, self.azimuth, -self.altitude);
-        let right = Y.cross(fwd_move);
-
-        let orient = orient_z(fwd, right).transpose();
-        let transl = translate(-self.pos);
-
-        transl * orient
     }
 }
